+++ diesel_tests/tests/lib.in.rs
@@ -1,2 +1,3 @@
 mod schema;
 mod insert;
+mod deserialization;