+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All user visible changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/), as described
 for Rust libraries in [RFC #1105](https://github.com/rust-lang/rfcs/blob/master/text/1105-api-evolution.md)
 
+## Unreleased
+
+### Fixed
+
+* `#[derive(Queriable)]` now allows generic parameters on the struct.
 
 ## [0.2.0] - 2015-11-30